+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -403,7 +403,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KR3GATGF6SPE5SPND0PWGBMR"),
+  _id: z.string().default("01KR3J2D5HREDER3SSP9H2G3FJ"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,