+++ packages/orq-rc/RELEASES.md
@@ -8949,3 +8949,13 @@ Based on:
 - [typescript v4.9.0-rc.33] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.33] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.33 - packages/orq-rc
+
+## 2026-05-08 10:30:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.34] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.34] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.34 - packages/orq-rc